schizo posted on the readme fiile
This is the extended description. I let the mind wander and typed the first thoughts in my head, much like now. It's all for the good of seeing how Github works so it doesn't really matter what I say as long as I understand what is happening. Barney the dinosaur is the sole survior becuase he hid underground with Godzilla. He realized instead of being feared he needed to be loved so he covered himself in felt to look more appealing and got a magic scientist to shrink him down to near human levels.

oh boy im sorry if anyone but me ever reads this